Move common Live Preview parts to PreviewBase
This lets us do the UJS page wiring in a consistent way via
componentDidMount, and move the stuff common to our two previews
into a single handleInputChange which is Rails-aware.
